refactor: 切り出したUpInsertのクラスを別ファイルに分離
